feat(ext-source): expose the last durable block to an external source
An external source that confirms progress upstream — advancing a Postgres replication slot, committing an offset — has had nothing to gate that on except a transaction's own durability handle, which completes once the transaction is replicated and applied. Object storage does not come into it. So a source acknowledging per-transaction is telling its upstream to discard data whose only copies are the replica log and the leader's live index, and the upstream cannot re-send it because we just said it needn't keep it.

`TxIndexer.latestBlock` is the position that is safe to confirm. It advances on `BlockCatalog.refresh`, which the leader calls once the block file has landed in storage and a follower once it has read the block back, so anything it emits is recoverable from storage alone.

A flow rather than a callback or a second per-transaction handle. A callback would put a network acknowledgement on the apply loop's critical path and make ordering the leader's problem rather than the source's; a handle per transaction would leave a block boundary sweeping a batch of them, where the confirmation position is monotonic and a conflated latest-wins value says the same thing once. The source already collects on its own coroutine inside `onPartitionAssigned`, so it has somewhere to put this.

`BlockDetails` moves out of `xtdb.catalog` in the process. Exposing it from `xtdb.api.tx` while `xtdb.catalog` imports `ExternalSourceToken` from that same package made a cycle, and the type is now part of the SPI rather than an implementation detail of the catalog. Its protobuf mapping deliberately stayed behind as a private extension: promoting the conversion too would have swapped the cycle for a block proto in `xtdb.api.tx`, which is the worse of the two. The move adds no new coupling — `BlockIndex` and `MessageId` are both typealiases for `Long`, and `DatabaseConfig` was already reachable from `xtdb.api` through `Storage`, `Log` and `LocalLog`.

Nothing consumes the flow yet; the Postgres source is gated separately. `ExternalSource`'s own documentation now separates confirming from resuming, because that distinction is the whole point — resume legitimately runs ahead of the last block, confirmation must not.

Refs #5867
