+++ core/src/main/kotlin/xtdb/api/tx/BlockDetails.kt
@@ -1,0 +1,27 @@
+package xtdb.api.tx
+
+import xtdb.api.TableRef
+import xtdb.api.TransactionKey
+import xtdb.database.proto.DatabaseConfig
+import xtdb.trie.BlockIndex
+import xtdb.types.MessageId
+
+/**
+ * What a block persisted to object storage carries.
+ *
+ * Deliberately knows nothing about the wire format: the block protobuf is the storage format, and the
+ * mapping from it lives with the catalog that reads it. Here a field a block may not carry is ordinary
+ * nullability, rather than a presence check at each read.
+ */
+data class BlockDetails(
+    val blockIndex: BlockIndex,
+    val latestCompletedTx: TransactionKey?,
+    val latestProcessedMsgId: MessageId?,
+    val boundaryReplicaMsgId: MessageId?,
+    // the leader term that produced this block's boundary; NONE for blocks written before
+    // term-fencing (see #5817)
+    val termId: Long,
+    val externalSourceToken: ExternalSourceToken?,
+    val tableNames: List<TableRef>,
+    val secondaryDatabases: Map<String, DatabaseConfig>,
+)
